CQE iterator
This adds a new API for processing CQEs as a psuedo-Iterator. Now, the APIs
which return multiple CQEs (ending with `for_cqes`) return a
`CompletionQueueEvents` type, which has a `next_cqe` and `for_each` method,
which allow the user to advance through the ready CQEs on the queue.

This API is designed to synchronize with the kernel as few times as possible
while processing all available CQEs. This is good if the code you perform for
each CQE is very fast, but it can lead to a backed-up CQ if that code takes
significant time. For example, this API would be good for a use which wakes
tasks when IO events complete (e.g. in the reactor of a Rust Futures API), but
not so much for something which actually completes the tasks synchronously on
the same thread the events are completed.

This commit also collapses some repetitious unsafe code into singular
underlying peek and wait functions, so that we should hopefully not see
divergences and bugs.